+++ packages/Adapter-ReactNative/src/index.ts
@@ -1,89 +1,94 @@
-import { InanoSQLAdapter, InanoSQLDataModel, InanoSQLTable, InanoSQLPlugin, InanoSQLInstance, VERSION } from "@nano-sql/core/lib/interfaces";
-import { generateID, setFast, deepSet } from "@nano-sql/core/lib/utilities";
+import { InanoSQLTable, InanoSQLPlugin, InanoSQLInstance } from "@nano-sql/core/lib/interfaces";
+import { generateID, deepSet, allAsync, binarySearch, chainAsync } from "@nano-sql/core/lib/utilities";
 import { nanoSQLMemoryIndex } from "@nano-sql/core/lib/adapters/memoryIndex";
-import * as mysql from "mysql";
+import { AsyncStorage } from "react-native";
 
-/*
-export interface mySQLConnection {
-    query: (sql: string, callback: (err: Error, results: any, fields: any) => void) => void;
-    release: () => void;
-}*/
+export const binaryInsert = (arr: any[], value: any, remove: boolean, startVal?: number, endVal?: number): boolean => {
 
-export class MySQL extends nanoSQLMemoryIndex {
+    const start = startVal || 0;
+    const end = endVal || arr.length;
+
+
+    if (arr[start] >= value) {
+        if (!remove) arr.unshift(value);
+        return remove ? false : true;
+    }
+    if (arr[end] <= value) {
+        if (!remove) arr.push(value);
+        return remove ? false : true;
+    }
+
+    const m = Math.floor((start + end) / 2);
+    if (value == arr[m]) { // already in array
+        if (remove) arr.splice(m, 1);
+        return remove ? true : false;
+    }
+    if (end - 1 == start) {
+        if (!remove) arr.splice(end, 0, value);
+        return remove ? false : true;
+    }
+
+    if (value > arr[m]) return binaryInsert(arr, value, remove, m, end);
+    if (value < arr[m]) return binaryInsert(arr, value, remove, start, m);
+
+    if (!remove) arr.splice(end, 0, value);
+    return remove ? false : true;
+};
+
+export class NativeStorage extends nanoSQLMemoryIndex {
 
     plugin: InanoSQLPlugin = {
-        name: "MySQL Adapter",
-        version: 2.01
+        name: "React Native Adapter",
+        version: 2.04
     };
 
     nSQL: InanoSQLInstance;
 
     private _id: string;
-    private _db: mysql.Pool;
+    private _indexes: {
+        [tableName: string]: any[];
+    }
+    private _ai: {
+        [tableName: string]: number;
+    }
     private _tableConfigs: {
         [tableName: string]: InanoSQLTable;
     }
 
-    constructor(public connectArgs: {
-        connectionLimit?: number;
-        host?: string;
-        port?: number;
-        socketPath?: string;
-        user: string;
-        password: string;
-        database: string;
-        charset?: string;
-        timezone?: string;
-        connectTimeout?: number;
-        stringifyObjects?: boolean;
-        insecureAuth?: boolean;
-        debug?: boolean;
-        trace?: boolean;
-        multipleStatements?: boolean;
-        ssl?: { [key: string]: any }
-    }) {
+    constructor(public cacheIndexes: boolean) {
         super(false, false);
         this._tableConfigs = {};
     }
 
     connect(id: string, complete: () => void, error: (err: any) => void) {
         this._id = id;
-        this._db = mysql.createPool({
-            connectionLimit: 20,
-            ...this.connectArgs
-        });
+        this._indexes = {};
+        this._ai = {};
+        complete();
+    }
 
-        this._db.getConnection((err, connection) => {
-            if (err) {
-                error(err);
+
+    key(table: string, pk: any) {
+        return this._id + "_" + table + "_" + String(pk);
+    }
+
+    getIndex(table: string): Promise<any[]> {
+        return new Promise((res, rej) => {
+            if (this._indexes[table]) {
+                res(this._indexes[table]);
                 return;
             }
-            connection.release();
-            complete();
-        });
-    }
-
-
-    private _chkTable(table: string): string {
-        if (Object.keys(this._tableConfigs).indexOf(table) === -1) {
-            throw Error("No table " + table + " found!");
-        } else {
-            return `\`${this._id}_${table}\``;
-        }
-    }
-
-    public _sql(sql: string, args: any[], complete: (rows: any) => void, error: (err: any) => void): void {
-        this._db.getConnection((err, db) => {
-            if (err) {
-                error(err); return;
-            };
-
-            db.query(mysql.format(sql, args), (err, result, fields) => {
+            AsyncStorage.getItem(this.key(table, "__IDX__"), (err, result) => {
                 if (err) {
-                    error(err); return;
-                };
-                db.release();
-                complete(result);
+                    rej(err);
+                    return;
+                }
+                if (this.cacheIndexes) {
+                    this._indexes[table] = JSON.parse(result || "[]");
+                    res(this._indexes[table]);
+                } else {
+                    res(JSON.parse(result || "[]"));
+                }
             });
         });
     }
@@ -91,111 +96,230 @@ export class MySQL extends nanoSQLMemoryIndex {
     createTable(tableName: string, tableData: InanoSQLTable, complete: () => void, error: (err: any) => void) {
 
         this._tableConfigs[tableName] = tableData;
-        this._sql(`CREATE TABLE IF NOT EXISTS ${this._chkTable(tableName)} (id ${tableData.isPkNum ? (tableData.pkType === "int" ? "INT" : "DOUBLE") : "VARCHAR(36)"} ${tableData.ai ? "AUTO_INCREMENT" : ""} PRIMARY KEY UNIQUE, data BLOB)`, [], (rows) => {
-            complete();
-        }, error);
-    }
-
-    dropTable(table: string, complete: () => void, error: (err: any) => void) {
-        this._sql(`DROP TABLE ${this._chkTable(table)};`, [], (rows) => {
-            delete this._tableConfigs[table];
-            complete();
-        }, error);
-    }
-
-    disconnect(complete: () => void, error: (err: any) => void) {
-        this._db.end((err) => {
+        AsyncStorage.getItem(this.key(tableName, "__IDX__"), (err, result) => {
             if (err) {
                 error(err);
                 return;
             }
-            complete();
+            if (result) {
+                AsyncStorage.getItem(this.key(tableName, "__AI__"), (err, ai) => {
+                    if (err) {
+                        error(err);
+                        return;
+                    }
+                    this._ai[tableName] = ai ? parseInt(ai) : 0;
+                    complete();
+                })
+            } else {
+                this._ai[tableName] = 0;
+                AsyncStorage.setItem(this.key(tableName, "__IDX__"), "[]", (err) => {
+                    if (err) {
+                        error(err);
+                        return;
+                    }
+                    AsyncStorage.setItem(this.key(tableName, "__AI__"), "0", (err) => {
+                        if (err) {
+                            error(err);
+                            return;
+                        }
+                        complete();
+                    })
+                })
+            }
         });
+
+    }
+
+    dropTable(table: string, complete: () => void, error: (err: any) => void) {
+        this.getIndex(table).then((pks) => {
+            AsyncStorage.multiRemove(pks.map(s => this.key(table, s)), (err) => {
+                if (err) {
+                    error(err);
+                    return;
+                }
+                AsyncStorage.removeItem(this.key(table, "__IDX__"), (err) => {
+                    if (err) {
+                        error(err);
+                        return;
+                    }
+                    AsyncStorage.removeItem(this.key(table, "__AI__"), (err) => {
+                        if (err) {
+                            error(err);
+                            return;
+                        }
+                        delete this._indexes[table];
+                        this._ai[table] = 0;
+                        complete();
+                    });
+                });
+            });
+        });
+    }
+
+    disconnect(complete: () => void, error: (err: any) => void) {
+        complete();
     }
 
     write(table: string, pk: any, row: { [key: string]: any }, complete: (pk: any) => void, error: (err: any) => void) {
 
-        pk = this._tableConfigs[table].ai ? pk : pk || generateID(this._tableConfigs[table].pkType, 0);
-
-        if (this._tableConfigs[table].ai && !pk) {
-            this._sql(`INSERT into ${this._chkTable(table)} (data) VALUES (?)`, [JSON.stringify(row)], (result) => {
-                deepSet(this._tableConfigs[table].pkCol, row, result.insertId);
-                this._sql(`UPDATE ${this._chkTable(table)} SET data = ? WHERE id = ?`, [JSON.stringify(row), result.insertId], () => {
-                    complete(result.insertId);
-                }, error);
-            }, error);
-        } else {
-            if (!pk) {
-                error("Can't add a row without a primary key!");
-                return;
-            }
-            const json = JSON.stringify(deepSet(this._tableConfigs[table].pkCol, row, pk));
-            this._sql(`INSERT into ${this._chkTable(table)} (id, data) VALUES (?, ?) ON DUPLICATE KEY UPDATE data = ?`, [pk, json, json], (result) => {
-                complete(pk);
-            }, error);
+        pk = pk || generateID(this._tableConfigs[table].pkType, this._ai[table] + 1);
+        if (typeof pk === "undefined") {
+            error(new Error("Can't add a row without a primary key!"));
+            return;
         }
+
+        if (this._tableConfigs[table].ai) {
+            this._ai[table] = Math.max(pk, this._ai[table]);
+        }
+
+        const json = JSON.stringify(deepSet(this._tableConfigs[table].pkCol, row, pk));
+        this.getIndex(table).then((index) => {
+            const didUpdate = binaryInsert(index, pk, false);
+            return allAsync([0, 1, 2], (idx, i, next, err) => {
+                switch(idx) {
+                    case 0:
+                        AsyncStorage.setItem(this.key(table, pk), json, (queryError) => {
+                            if (queryError) {
+                                error(queryError);
+                                return;
+                            }
+                            next();
+                        });
+                    break;
+                    case 1:
+                        if (didUpdate) {
+                            AsyncStorage.setItem(this.key(table, "__IDX__"), JSON.stringify(index), (queryError) => {
+                                if (queryError) {
+                                    err(queryError);
+                                    return;
+                                }
+                                next();
+                            })
+                        } else {
+                            next();
+                        }
+                    break;
+                    case 2:
+                        if (this._tableConfigs[table].ai) {
+                            AsyncStorage.setItem(this.key(table, "__AI__"), String(this._ai[table]), (queryError) => {
+                                if (queryError) {
+                                    err(queryError);
+                                    return;
+                                }
+                                next();
+                            })
+                        } else {
+                            next();
+                        }
+                    break;
+                }
+            });
+        }).then(() => {
+            complete(pk);
+        }).catch(error);
+
     }
 
     read(table: string, pk: any, complete: (row: { [key: string]: any } | undefined) => void, error: (err: any) => void) {
-        this._sql(`SELECT * FROM ${this._chkTable(table)} WHERE id = ?`, [pk], (rows) => {
-            if (rows.length) {
-                complete(JSON.parse(rows[0].data.toString('utf8')));
-            } else {
-                complete(undefined);
+        AsyncStorage.getItem(this.key(table, pk), (err, result) => {
+            if (err) {
+                error(err);
+                return;
             }
-        }, error);
+            complete(result ? JSON.parse(result) : undefined);
+        });
     }
 
     readMulti(table: string, type: "range" | "offset" | "all", offsetOrLow: any, limitOrHigh: any, reverse: boolean, onRow: (row: { [key: string]: any }, i: number) => void, complete: () => void, error: (err: any) => void) {
 
-        let stmnt = `SELECT data FROM ${this._chkTable(table)}`;
+        let range: any[] = {
+            "range":  [offsetOrLow, limitOrHigh],
+            "offset": [offsetOrLow, offsetOrLow + limitOrHigh],
+            "all": []
+        }[type];
 
-        if (type === "range") {
-            stmnt += ` WHERE id BETWEEN ? AND ?`;
-        }
-        if (reverse) {
-            stmnt += ` ORDER BY id DESC`;
-        } else {
-            stmnt += ` ORDER BY id`;
-        }
 
-        let query = stmnt;
-        if (type === "offset") {
-            const lower = reverse ? offsetOrLow + 1 : offsetOrLow;
-            const higher = limitOrHigh;
-            query += ` LIMIT ${higher} OFFSET ${lower}`;
-        }
-        
-        let count = 0;
-        this._db.query(query, type === "range" ? [offsetOrLow, limitOrHigh] : []).on('error', function (err) {
-            error(err);
-        }).on('result', function (row) {
-            onRow(JSON.parse(row.data.toString('utf8')), count);
-            count++;
-        }).on('end', function () {
+        this.getIndex(table).then((index) => {
+            return new Promise((res, rej) => {
+                switch(type) {
+                    case "all":
+                        res(index.slice());
+                        break;
+                    case "offset":
+                        const l = index.length - 1;
+                        res(reverse ? index.slice(l - range[1], l - range[0]) : index.slice(range[0], range[1]));
+                        break;
+                    case "range":
+                        let lowIdx = binarySearch(index, range[0], false);
+                        let highIdx = binarySearch(index, range[1], false);
+    
+                        while(index[highIdx] > range[1]) {
+                            highIdx--;
+                        }
+    
+                        while(index[lowIdx] < range[0]) {
+                            lowIdx++;
+                        }
+                        res(index.slice(lowIdx, highIdx + 1));
+                        break;
+                    default:
+                        res([]);
+                }
+            });
+        }).then((getPKs: any[]) => {
+            return chainAsync(reverse ? getPKs.reverse() : getPKs, (pk, i, next, err) => {
+                this.read(table, pk, (row) => {
+                    onRow(row || {}, i);
+                    next();
+                }, err);
+            });
+        }).then(() => {
             complete();
-        });
+        }).catch(error);
+
     }
 
     delete(table: string, pk: any, complete: () => void, error: (err: any) => void) {
-        this._sql(`DELETE FROM ${this._chkTable(table)} WHERE id = ?`, [pk], () => {
-            complete();
-        }, error);
+        this.getIndex(table).then((index) => {
+            const didUpdate = binaryInsert(index, pk, true);
+            allAsync([0, 1], (item, i, next, err) => {
+                switch(item) {
+                    case 0:
+                        if (didUpdate) {
+                            AsyncStorage.setItem(this.key(table, "__IDX__"), JSON.stringify(index), (queryError) => {
+                                if (queryError) {
+                                    err(queryError);
+                                    return;
+                                }
+                                next();
+                            })
+                        } else {
+                            next();
+                        }
+                    break;
+                    case 1:
+                        AsyncStorage.removeItem(this.key(table, pk), (queryError) => {
+                            if (queryError) {
+                                err(queryError);
+                                return;
+                            }
+                            next();
+                        });
+                    break;
+                }
+            }).then(() => {
+                complete();
+            }).catch(error);
+        }).catch(error);
     }
 
     getTableIndex(table: string, complete: (index: any[]) => void, error: (err: any) => void) {
-        this._sql(`SELECT id FROM ${this._chkTable(table)} ORDER BY id`, [], (rows) => {
-            let idx: any[] = [];
-            for (let i = 0; i < rows.length; i++) {
-                idx.push(rows[i].id);
-            }
-            complete(idx);
-        }, error);
+        this.getIndex(table).then(complete).catch(error);
     }
 
     getTableIndexLength(table: string, complete: (length: number) => void, error: (err: any) => void) {
-        this._sql(`SELECT COUNT(*) FROM ${this._chkTable(table)}`, [], (rows) => {
-            complete(rows[0]["COUNT(*)"]);
-        }, error);
+        this.getIndex(table).then((index) => {
+            complete(index.length);
+        }).catch(error);
     }
 }